requirements.txt
streamlit
azure-cognitiveservices-speech
python-dotenv
streamlit_webrtc
<<<<<<< HEAD
=======
vinorm
langdetect
underthesea
re
datetime
>>>>>>> a3cbcbf (Your commit message here)
